[Dy2St][CINN] Dont get cpp flag `FLAGS_use_cinn` by `os.getenv`
